Create README.md
Pillow Palooza is a real estate start-up. The main objective is to get insights on short term rental 
market in New York City. This analysis is based on Airbnb New York’s nine months data. The insights 
gained from the analysis will help Pillow Palooza to enhance its revenue and occupancy rates for their 
short-term rental properties. For the analysis three techniques were used to gain insights. Data 
Wrangling and cleaning is done with Python. Data Analysis and insights were generated with the help 
of SQL and visualizations with Tableau